Split Name::Scopes tests out of name_test.rb (#4708)
Moves the 11 tests exercising Name::Scopes (matching_desired_new_parsed_name,
subtaxa_of, classification_has, species_lists, within_locations, in_box,
show_includes/merge_includes) into test/models/name/scopes_test.rb,
matching app/models/name/scopes.rb.

test_show_includes_omits_namings_and_observations and
test_merge_includes_preloads_namings_and_observations lived after the
old file's Validations section (added in an earlier commit here) with
no section header of their own - moved along with the rest since both
exercise scopes defined in scopes.rb.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
